fix user service namespace
