+++ app/graphql/mutations/users/login.rb
@@ -16,7 +16,7 @@ module Mutations
       require_one_of %i[email username]
 
       def resolve(args)
-        response = UserLoginService.new(args).execute.to_mutation_response(success_key: :user_session)
+        response = ::Users::LoginService.new(args).execute.to_mutation_response(success_key: :user_session)
         bypass_authorization! response, object_path: %i[user_session user]
         bypass_authorization! response, object_path: :user_session
       end
