+++ app/graphql/mutations/users/register.rb
@@ -14,7 +14,7 @@ module Mutations
       argument :username, String, required: true, description: 'Username of the user'
 
       def resolve(username:, email:, password:)
-        response = UserRegisterService.new(username, email, password).execute.to_mutation_response(success_key: :user)
+        response = ::Users::RegisterService.new(username, email, password).execute.to_mutation_response(success_key: :user)
         bypass_authorization! response, object_path: :user
       end
     end
